Consistently use size_t and SAL_MAX_SIZE
Instead of a mix of sal_uIntPtr, sal_uLong, int, and so on.
Also change CONTAINER_ENTRY_NOTFOUND=ULONG_MAX to SAL_MAX_SIZE as
return value in case of failure and in the related tests.

+++ svx/source/tbxctrls/grafctrl.cxx
@@ -871,7 +871,7 @@ void SvxGrafAttrHelper::GetGrafAttrState( SfxItemSet& rSet, SdrView& rView )
     bool            bEnableTransparency = true;
     bool            bEnableCrop = ( 1 == rMarkList.GetMarkCount() );
 
-    for( int i = 0, nCount = rMarkList.GetMarkCount(); i < nCount; ++i )
+    for( size_t i = 0, nCount = rMarkList.GetMarkCount(); i < nCount; ++i )
     {
         SdrGrafObj* pGrafObj = dynamic_cast< SdrGrafObj* >( rMarkList.GetMark( i )->GetMarkedSdrObj() );
